logic: Hide non-fatal errors in unindexed files
* added database field indexed to error to tell if the error occured within an indexed file
* added checkbox to Preferences for showing them

+++ src/lib/data/ErrorCountInfo.h
@@ -8,12 +8,12 @@ struct ErrorCountInfo
 		, fatal(0)
 	{}
 
-	ErrorCountInfo(int total, size_t fatal)
+	ErrorCountInfo(size_t total, size_t fatal)
 		: total(total)
 		, fatal(fatal)
 	{}
 
-	int total;
+	size_t total;
 	size_t fatal;
 };
